+++ packages/atomic/src/components.d.ts
@@ -7,6 +7,7 @@
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { CategoryFacetSortCriterion, FacetSortCriterion, LogLevel, NumericFilter, NumericFilterState, RangeFacetRangeAlgorithm, RangeFacetSortCriterion, Result, ResultTemplate, ResultTemplateCondition, SearchEngine } from "@coveo/headless";
 import { Bindings } from "./utils/initialization-utils";
+import { NumberInputType } from "./components/facets-v1/facet-number-input/number-input-type";
 import { i18n } from "i18next";
 import { InitializationOptions } from "./components/atomic-search-interface/atomic-search-interface";
 export namespace Components {
@@ -137,7 +138,7 @@ export namespace Components {
         "filter": NumericFilter;
         "filterState": NumericFilterState;
         "label": string;
-        "onApply"?: () => void;
+        "type": NumberInputType;
     }
     interface AtomicFacetV1 {
         /**
@@ -276,7 +277,7 @@ export namespace Components {
          */
         "label": string;
         /**
-          * The number of values to request for this facet, when there are no manual ranges.
+          * The number of values to request for this facet, when there are no manual ranges. If the number of values is 0, no ranges will be displayed.
          */
         "numberOfValues": number;
         /**
@@ -288,13 +289,9 @@ export namespace Components {
          */
         "sortCriteria": RangeFacetSortCriterion;
         /**
-          * Whether this facet should contain an input allowing users to set custom ranges.
+          * Whether this facet should contain an input allowing users to set custom ranges. Depending on the field, the input can allow either decimal or integer values.
          */
-        "withInput": boolean;
-        /**
-          * Whether this facet should contain a defined list of values.
-         */
-        "withRanges": boolean;
+        "withInput"?: NumberInputType;
     }
     interface AtomicNumericRange {
         /**
@@ -346,11 +343,11 @@ export namespace Components {
          */
         "label": string;
         /**
-          * The maximum value of the field. This value is used to normalize the field values with the number of stars. Default is 5.
+          * The maximum value of the field. This value is used to normalize the field values with the number of stars.
          */
-        "maxValue": number;
+        "maxValueInIndex": number;
         /**
-          * The number of stars to request for this facet. Default is 5.
+          * The number of stars to request for this facet.
          */
         "numberOfStars": number;
     }
@@ -1036,7 +1033,8 @@ declare namespace LocalJSX {
         "filter": NumericFilter;
         "filterState": NumericFilterState;
         "label": string;
-        "onApply"?: () => void;
+        "onAtomic/numberInputApply"?: (event: CustomEvent<any>) => void;
+        "type": NumberInputType;
     }
     interface AtomicFacetV1 {
         /**
@@ -1175,7 +1173,7 @@ declare namespace LocalJSX {
          */
         "label"?: string;
         /**
-          * The number of values to request for this facet, when there are no manual ranges.
+          * The number of values to request for this facet, when there are no manual ranges. If the number of values is 0, no ranges will be displayed.
          */
         "numberOfValues"?: number;
         /**
@@ -1187,13 +1185,9 @@ declare namespace LocalJSX {
          */
         "sortCriteria"?: RangeFacetSortCriterion;
         /**
-          * Whether this facet should contain an input allowing users to set custom ranges.
+          * Whether this facet should contain an input allowing users to set custom ranges. Depending on the field, the input can allow either decimal or integer values.
          */
-        "withInput"?: boolean;
-        /**
-          * Whether this facet should contain a defined list of values.
-         */
-        "withRanges"?: boolean;
+        "withInput"?: NumberInputType;
     }
     interface AtomicNumericRange {
         /**
@@ -1246,11 +1240,11 @@ declare namespace LocalJSX {
          */
         "label"?: string;
         /**
-          * The maximum value of the field. This value is used to normalize the field values with the number of stars. Default is 5.
+          * The maximum value of the field. This value is used to normalize the field values with the number of stars.
          */
-        "maxValue"?: number;
+        "maxValueInIndex"?: number;
         /**
-          * The number of stars to request for this facet. Default is 5.
+          * The number of stars to request for this facet.
          */
         "numberOfStars"?: number;
     }