+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.40.1...@coveo/atomic@0.40.2) (2021-07-13)
+
+
+### Bug Fixes
+
+* **atomic:** various numeric facet v1 fixes & improvements ([#958](https://github.com/coveo/ui-kit/issues/958)) ([8fb78a2](https://github.com/coveo/ui-kit/commit/8fb78a21f6d9197aa694981c67295ba7f942b8e2))
+
+
+
+
+
+## [0.40.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.40.0...@coveo/atomic@0.40.1) (2021-07-13)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.40.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.39.3...@coveo/atomic@0.40.0) (2021-07-09)
+
+
+### Features
+
+* **atomic:** skeleton for Atomic-Rating-Facet ([#957](https://github.com/coveo/ui-kit/issues/957)) ([0639e63](https://github.com/coveo/ui-kit/commit/0639e632c9d2169ad1a6b88761bd2b5c47fde8d2))
+
+
+
+
+
 ## [0.39.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.39.2...@coveo/atomic@0.39.3) (2021-07-09)
 
 **Note:** Version bump only for package @coveo/atomic